some clean up and cache
