+++ kamon-annotation/src/main/java/kamon/annotation/instrumentation/advisor/HistogramAnnotationAdvisor.java
@@ -16,6 +16,7 @@
 
 package kamon.annotation.instrumentation.advisor;
 
+import kamon.annotation.instrumentation.cache.AnnotationCache;
 import kamon.metric.Histogram;
 import kanela.agent.libs.net.bytebuddy.asm.Advice;
