+++ kamon-annotation/src/main/java/kamon/annotation/instrumentation/advisor/RangeSamplerAnnotationAdvisor.java
@@ -16,6 +16,7 @@
 
 package kamon.annotation.instrumentation.advisor;
 
+import kamon.annotation.instrumentation.cache.AnnotationCache;
 import kamon.metric.RangeSampler;
 import kanela.agent.libs.net.bytebuddy.asm.Advice;
