+++ kamon-annotation/src/main/resources/reference.conf
@@ -18,15 +18,4 @@ kanela {
       ]
     }
   }
-
-  class-dumper {
-    # Enables the class dumper.
-    enabled = true
-    # Default dir.
-    dir = ${user.home}"/kanela-agent/dump"
-    # By default build a jar containing all instrumented classes.
-    create-jar = true
-    # Default jar name.
-    jar-name = "instrumented-classes-annotations"
-  }
 }
